Add files via upload
This program is for a beverage vending machine that dispenses two kinds of beers i.e tusker and guinnes.  The vending unit utilizes a single dispensing unit that dispenses from the two storage sections. When a servo motor is rotated clockwise, the tusker brand is dispensed and when rotated counter clockwise, the guinnes brand is dispensed.